+++ mitra_activitypub/src/builders/mod.rs
@@ -23,5 +23,6 @@ pub mod remove_person;
 pub mod undo_announce;
 pub mod undo_follow;
 pub mod undo_like;
+pub mod update_agreement;
 pub mod update_note;
 pub mod update_person;